part-dataeng/fastapi_pubsub_requirements.txt
absl-py==1.0.0
anyio==3.6.1
asgiref==3.5.2
asn1crypto==0.24.0
astunparse==1.6.3
attrs==17.4.0
Automat==0.6.0
blinker==1.4
cachetools==5.1.0
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloud-init==22.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
fastapi==0.78.0
flatbuffers==1.12
gast==0.4.0
google-api-core==2.8.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-pubsub==2.12.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
h5py==3.6.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.3
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==1.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
PAM==0.4.2
pexpect==4.2.1
Pillow==9.1.1
proto-plus==1.20.4
protobuf==3.20.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pydantic==1.9.1
PyGObject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==3.0.9
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
PyWavelets==1.3.0
pyxdg==0.25
PyYAML==3.12
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.8
scikit-image==0.19.2
scipy==1.8.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sniffio==1.2.0
sos==4.3
ssh-import-id==5.7
starlette==0.19.1
systemd-python==234
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.5.4
Twisted==17.9.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.7
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.3
urllib3==1.22
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
zope.interface==4.3.2
